Exercise-Observing a TextField - SwiftUI
In this exercise, you will practice working with TextFields. You will also learn how to store user input in control variables.
Scenario
Suppose you are creating an app for the Little Lemon restaurant that allows customers to make reservations.  The app must prompt the user to enter their name during one of the steps of making a reservation.